rename test &&add one
